requirements.txt
eth-ape'[recommended-plugins]'==0.6.15